+++ netwerk/test/TestRes.cpp
@@ -150,7 +150,7 @@ public:
 
     NS_IMETHOD OnDataAvailable(nsIRequest *request, nsISupports *ctxt, 
                                nsIInputStream *inStr,
-                               uint32_t sourceOffset, uint32_t count) {
+                               uint64_t sourceOffset, uint32_t count) {
         nsresult rv;
         char buf[1024];
         while (count > 0) {
